fix: tsc, lint and packages
