+++ monorepo/packages/governance/package.json
@@ -24,11 +24,7 @@
     "@storybook/react": "7.0.10",
     "@storybook/react-webpack5": "7.0.10",
     "@types/react": "^18.2.5",
-    "@types/react-custom-scrollbars": "4.0.10",
     "@types/styled-components": "5.1.15",
-    "@types/react-window": "^1.8.2",
-    "@types/numeral": "^2.0.2",
-    "@types/prop-types": "^15.7.4",
     "eslint": "^7.32.0",
     "react": "^18.2.0",
     "react-dom": "18.2.0",
@@ -44,27 +40,21 @@
     "typescript": "^4.5.2"
   },
   "dependencies": {
+    "@ethersproject/abi": "5.6.4",
     "@ethersproject/providers": "5.7.2",
+    "@pangolindex/exchange-contracts": "2.1.7",
     "@pangolindex/governance": "^1.2.0",
     "@pangolindex/shared": "*",
     "@pangolindex/core": "*",
     "@pangolindex/sar": "*",
     "@pangolindex/state-hooks": "*",
     "@pangolindex/wallet-connectors": "*",
-    "axios": "0.26.1",
-    "dayjs": "1.10.7",
-    "jotai": "2.0.3",
-    "lodash": "4.0.1",
-    "numeral": "2.0.6",
-    "prop-types": "15.7.2",
-    "qs": "6.10.3",
+    "ethers": "^5.5.4",
+    "luxon": "1.28.1",
+    "react-markdown": "^8.0.0",
+    "remark-gfm": "^3.0.1",
     "styled-components": "5.3.3",
-    "react-custom-scrollbars": "4.2.1",
-    "react-feather": "2.0.9",
-    "react-window": "1.8.5",
-    "react-query": "3.39.0",
-    "react-use": "15.3.8",
-    "recharts": "2.1.13"
+    "react-feather": "2.0.9"
   },
   "peerDependencies": {
     "react": ">=17.0.2",
