+++ monorepo/packages/governance/src/hooks/evm.ts
@@ -1,5 +1,4 @@
 import { TransactionResponse } from '@ethersproject/providers';
-
 import { ChainId, TokenAmount } from '@pangolindex/sdk';
 import {
   GET_BLOCK,
